// src/app/train/model.interface.ts
export interface IModel extends IAsset {
  status: IModel.Status;
  dataType?: string;
  tag?: string;
  size?: number;
  filePath?: string;
  experimentId?: TObjectId;
}

export interface ITabularModel extends IModel {
  responseColumn: IModelColumn;
  predictorColumns: IModelColumn[];
  class: ITabularModel.Class;
  experimentId?: TObjectId;
  classes?: String[];
  inLibrary?: boolean;
}

export interface IModelUpdate {
  name?: string;
  description?: string;
}

export interface IModelCreate {
  name: string;
  dataType?: string;
  size?: number;
  filePath?: string;
  experimentId?: TObjectId;
}

export interface IModelImport {
  name: string;
}

export interface IModelClone {
  name: string;
  description?: string;
}

// retainined "training" and "predicting" for old tabular models - see note above
export namespace IModel {
  export enum Status {
    ACTIVE = 'ACTIVE',
    IMPORTING = 'IMPORTING',
    ERROR = 'ERROR',
    CANCELLED = 'CANCELLED',
  }
}